add first py file

+++ main.py
@@ -1,2 +1,2 @@
-def add(a: int, b: int) -> float:
+def add(a: int, b: int) -> int:
     return a + b
